requirements.txt
backports==1.0
backports.functools_lru_cache==1.6.4
blas==1.0
bottleneck==1.3.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2022.10.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
cryptography==38.0.1
cudatoolkit==11.3.1
cycler==0.11.0
dbus==1.13.18
et_xmlfile==1.1.0
expat==2.4.9
ffmpeg==4.3
fftw==3.3.9
fontconfig==2.13.1
fonttools==4.25.0
freetype==2.11.0
ftfy==6.1.1
gettext==0.21.0
giflib==5.2.1
glib==2.69.1
gmp==6.2.1
gnutls==3.6.15
gst-plugins-base==1.14.0
gstreamer==1.14.0
icu==58.2
idna==3.4
intel-openmp==2021.4.0
joblib==1.2.0
jpeg==9e
kiwisolver==1.4.2
krb5==1.19.2
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libclang==10.0.1
libcxx==14.0.6
libcxxabi==14.0.6
libdeflate==1.8
libedit==3.1.20210910
libevent==2.1.12
libffi==3.3
libgcc-ng==12.2.0
libgfortran-ng==11.3.0
libgfortran5==11.3.0
libgomp==12.2.0
libiconv==1.16
libidn2==2.3.2
liblapack==3.9.0
libllvm10==10.0.1
libpng==1.6.37
libpq==12.9
libsqlite==3.39.3
libstdcxx-ng==12.2.0
libtasn1==4.16.0
libtiff==4.4.0
libunistring==0.9.10
libuuid==1.0.3
libwebp==1.2.4
libwebp-base==1.2.4
libxcb==1.15
libxkbcommon==1.0.1
libxml2==2.9.14
libxslt==1.1.35
libzlib==1.2.13
llvm-openmp==14.0.6
lz4-c==1.9.3
matplotlib==3.5.1
matplotlib-base==3.5.1
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
munkres==1.1.4
ncurses==6.3
nettle==3.7.3
nltk==3.7
nspr==4.33
nss==3.74
numexpr==2.8.3
numpy==1.23.3
numpy-base==1.23.3
openh264==2.1.1
openpyxl==3.0.10
openssl==1.1.1s
packaging==21.3
pandas==1.4.4
pcre==8.45
pillow==9.2.0
pip==22.2.2
ply==3.11
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.9
pyqt==5.15.7
pyqt5-sip==12.11.0
pysocks==1.7.1
python==3.9.13
python-dateutil==2.8.2
python_abi==3.9
pytorch==1.12.1
pytorch-mutex==1.0
pytz==2022.1
qt-main==5.15.2
qt-webengine==5.15.9
qtwebkit==5.212
readline==8.2
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.1
scipy==1.9.1
seaborn==0.11.2
setuptools==63.4.1
sip==6.6.2
six==1.16.0
sqlite==3.39.3
threadpoolctl==3.1.0
tk==8.6.12
toml==0.10.2
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022e
urllib3==1.26.12
wcwidth==0.2.5
wheel==0.37.1
xz==5.2.6
zlib==1.2.13
zstd==1.5.2
